CursorWorkplace/Quickbook/requirements.txt
quickbooks-python==0.9.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
lxml==4.9.3 